+++ pkgs/bazel_worker/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0
+
+* Improve `AsyncMessageGrouper` performance.
+* Add `benchmark/benchmark.dart` measuring `AsyncMessageGrouper` performance.
+
 ## 1.0.0-nullsafety.0
 
 * Migrate to null safety.